commenti e piccole modifiche

+++ scripts/deploy.sh
@@ -1,6 +1,7 @@
 PROJECT_FOLDER="$DNSServerApp"
 DB_NAME="$staging.sqlite3"
 ENV_FILE="$.env.staging"
+cd ~/
 if [ ! -d $PROJECT_FOLDER ] ; then
 echo 'NOT Exists'
 else
@@ -32,10 +33,17 @@ cd ~/
 if [ -f staging.sqlite3 ] ; then
 echo 'Move DB'
 mv ~/staging.sqlite3 ~/DNSServerApp/backend/src/db/staging.sqlite3
+else
+echo 'missing DB file, will create empty one'
 fi
 
-
-mv ~/.env.staging ~/DNSServerApp/backend/env/.env.staging
+if [ -f .env.staging ] ; then
+ echo 'Move env file'
+ mv ~/.env.staging ~/DNSServerApp/backend/env/.env.staging
+else
+ echo 'missing env file, please create one, using example.env. EXITING'
+ exit 1 
+fi
 
 cd ~/DNSServerApp/backend;
 NODE_ENV=staging pm2 start dist/main.js --name "dnsserverapp" && cd ~/ && pm2 startup > pm2_startup_output;
